Move most `Identify*` functions to their respective files
Can't move `IdentifyShopItem` as it's used all over the place.
