+++ soh/soh/Enhancements/randomizer/ShuffleRedIce.cpp
@@ -3,6 +3,7 @@
 #include "item_category_adj.h"
 #include "particle_cmc.h"
 #include "soh/Enhancements/randomizer/randomizer.h"
+#include "soh/Enhancements/randomizer/RCToRandInf.h"
 
 extern "C" {
 #include "functions.h"
@@ -118,6 +119,33 @@ void BgIceShelter_RandomizerSpawnCollectible(Actor* actor) {
     }
 }
 
+static CheckIdentity IdentifyRedIce(s32 sceneNum, s32 posX, s32 posZ) {
+    struct CheckIdentity redIceIdentity;
+    uint32_t redIceSceneNum = sceneNum;
+
+    // Handle KZ moving
+    if (sceneNum == SCENE_ZORAS_DOMAIN && LINK_IS_ADULT && posX == 531 && posZ == -1818) {
+        posX = 628;
+    }
+
+    redIceIdentity.randomizerInf = RAND_INF_MAX;
+    redIceIdentity.randomizerCheck = RC_UNKNOWN_CHECK;
+
+    s32 actorParams = TWO_ACTOR_PARAMS(posX, posZ);
+
+    Rando::Location* location = OTRGlobals::Instance->gRandomizer->GetCheckObjectFromActor(ACTOR_BG_ICE_SHELTER, redIceSceneNum, actorParams);
+
+    if (location->GetRandomizerCheck() == RC_UNKNOWN_CHECK) {
+        LUSLOG_WARN("IdentifyRedIce did not receive a valid RC value (%d).", location->GetRandomizerCheck());
+        assert(false);
+    } else {
+        redIceIdentity.randomizerInf = rcToRandomizerInf[location->GetRandomizerCheck()];
+        redIceIdentity.randomizerCheck = location->GetRandomizerCheck();
+    }
+
+    return redIceIdentity;
+}
+
 void RegisterShuffleRedIce() {
     bool shouldRegister = IS_RANDO && Rando::Context::GetInstance()->GetOption(RSK_SHUFFLE_RED_ICE).Get();
 
@@ -125,7 +153,7 @@ void RegisterShuffleRedIce() {
         BgIceShelter* redIceActor = va_arg(args, BgIceShelter*);
         Actor* actor = (Actor*)redIceActor;
 
-        auto redIceIdentity = OTRGlobals::Instance->gRandomizer->IdentifyRedIce(
+        auto redIceIdentity = IdentifyRedIce(
             gPlayState->sceneNum, (s16)actor->world.pos.x, (s16)actor->world.pos.z);
         ObjectExtension::GetInstance().Set<CheckIdentity>(actor, std::move(redIceIdentity));
