+++ soh/soh/Enhancements/randomizer/ShuffleSigns.cpp
@@ -3,6 +3,7 @@
 #include "item_category_adj.h"
 #include "particle_cmc.h"
 #include "soh/Enhancements/randomizer/randomizer.h"
+#include "soh/Enhancements/randomizer/RCToRandInf.h"
 
 extern "C" {
 extern PlayState* gPlayState;
@@ -89,6 +90,65 @@ void Sign_RoyalTombSpawnCollectible(int16_t flagType, int16_t flag) {
     }
 }
 
+static CheckIdentity IdentifySign(s32 sceneNum, s32 posX, s32 posZ, s32 id) {
+    CheckIdentity signIdentity;
+    uint32_t signSceneNum = sceneNum;
+    Rando::Location* location = nullptr;
+
+    // align child/adult signs
+    if (sceneNum == SCENE_KAKARIKO_VILLAGE && LINK_IS_ADULT && posX == 1165 && posZ == 1545) {
+        posZ = 1550;
+    } else if (sceneNum == SCENE_GRAVEYARD && LINK_IS_ADULT) {
+        if (id == ACTOR_EN_WONDER_TALK2 && posX == -807 && posZ == 266) {
+            posX = -805;
+        } else if (id == ACTOR_EN_WONDER_TALK) {
+            if (posX == 634 && posZ == 260) {
+                posX = 654;
+                posZ = 258;
+            } else if (posX == 634 && posZ == -100) {
+                posX = 654;
+                posZ = -102;
+            } else if (posX == 753 && posZ == 85) {
+                posX = 752;
+            }
+        }
+    } else if (sceneNum == SCENE_ZORAS_RIVER && LINK_IS_ADULT && posX == 4097 && posZ == -1399) {
+        posX = 4096;
+        posZ = -1401;
+    }
+
+    signIdentity.randomizerInf = RAND_INF_MAX;
+    signIdentity.randomizerCheck = RC_UNKNOWN_CHECK;
+
+    s32 actorParams = TWO_ACTOR_PARAMS(posX, posZ);
+
+    switch (id) {
+        case ACTOR_EN_KANBAN:
+            location = OTRGlobals::Instance->gRandomizer->GetCheckObjectFromActor(ACTOR_EN_KANBAN, signSceneNum, actorParams);
+            break;
+        case ACTOR_EN_A_OBJ:
+            location = OTRGlobals::Instance->gRandomizer->GetCheckObjectFromActor(ACTOR_EN_A_OBJ, signSceneNum, actorParams);
+            break;
+        case ACTOR_EN_WONDER_TALK2:
+            location = OTRGlobals::Instance->gRandomizer->GetCheckObjectFromActor(ACTOR_EN_WONDER_TALK2, signSceneNum, actorParams);
+            break;
+        case ACTOR_EN_WONDER_TALK:
+            location = OTRGlobals::Instance->gRandomizer->GetCheckObjectFromActor(ACTOR_EN_WONDER_TALK, signSceneNum, actorParams);
+            break;
+        default:
+            return signIdentity;
+    }
+
+    if (location == nullptr || location->GetRandomizerCheck() == RC_UNKNOWN_CHECK) {
+        LUSLOG_WARN("IdentifySign did not receive a valid RC value (%d).", location->GetRandomizerCheck());
+    } else {
+        signIdentity.randomizerInf = rcToRandomizerInf[location->GetRandomizerCheck()];
+        signIdentity.randomizerCheck = location->GetRandomizerCheck();
+    }
+
+    return signIdentity;
+}
+
 void RegisterShuffleSigns() {
     bool shouldRegister = IS_RANDO && Rando::Context::GetInstance()->GetOption(RSK_SHUFFLE_SIGNS).Get();
 
@@ -96,7 +156,7 @@ void RegisterShuffleSigns() {
         Actor* actor = static_cast<Actor*>(actorRef);
         EnKanban* signActor = static_cast<EnKanban*>(actorRef);
 
-        auto signIdentity = OTRGlobals::Instance->gRandomizer->IdentifySign(
+        auto signIdentity = IdentifySign(
             gPlayState->sceneNum, (s16)actor->world.pos.x, (s16)actor->world.pos.z, actor->id);
         ObjectExtension::GetInstance().Set<CheckIdentity>(actor, std::move(signIdentity));
     });
@@ -105,7 +165,7 @@ void RegisterShuffleSigns() {
         Actor* actor = static_cast<Actor*>(actorRef);
         EnAObj* signActor = static_cast<EnAObj*>(actorRef);
 
-        auto signIdentity = OTRGlobals::Instance->gRandomizer->IdentifySign(
+        auto signIdentity = IdentifySign(
             gPlayState->sceneNum, (s16)actor->world.pos.x, (s16)actor->world.pos.z, actor->id);
         ObjectExtension::GetInstance().Set<CheckIdentity>(actor, std::move(signIdentity));
     });
@@ -114,7 +174,7 @@ void RegisterShuffleSigns() {
         Actor* actor = static_cast<Actor*>(actorRef);
         EnWonderTalk* signActor = static_cast<EnWonderTalk*>(actorRef);
 
-        auto signIdentity = OTRGlobals::Instance->gRandomizer->IdentifySign(
+        auto signIdentity = IdentifySign(
             gPlayState->sceneNum, (s16)actor->world.pos.x, (s16)actor->world.pos.z, actor->id);
         ObjectExtension::GetInstance().Set<CheckIdentity>(actor, std::move(signIdentity));
     });
@@ -123,7 +183,7 @@ void RegisterShuffleSigns() {
         Actor* actor = static_cast<Actor*>(actorRef);
         EnWonderTalk2* signActor = static_cast<EnWonderTalk2*>(actorRef);
 
-        auto signIdentity = OTRGlobals::Instance->gRandomizer->IdentifySign(
+        auto signIdentity = IdentifySign(
             gPlayState->sceneNum, (s16)actor->world.pos.x, (s16)actor->world.pos.z, actor->id);
         ObjectExtension::GetInstance().Set<CheckIdentity>(actor, std::move(signIdentity));
     });
