+++ soh/soh/Enhancements/randomizer/ShuffleTrees.cpp
@@ -4,6 +4,7 @@
 #include "soh/ObjectExtension/ObjectExtension.h"
 #include "item_category_adj.h"
 #include "soh/Enhancements/randomizer/randomizer.h"
+#include "soh/Enhancements/randomizer/RCToRandInf.h"
 
 extern "C" {
 #include "variables.h"
@@ -125,6 +126,27 @@ void EnWood02_RandomizerSpawnCollectible(EnWood02* treeActor, PlayState* play) {
     treeIdentity->randomizerCheck = RC_UNKNOWN_CHECK;
 }
 
+static CheckIdentity IdentifyTree(s32 sceneNum, s32 posX, s32 posZ) {
+    CheckIdentity treeIdentity;
+
+    if (sceneNum == SCENE_MARKET_NIGHT) {
+        sceneNum = SCENE_MARKET_DAY;
+    }
+
+    s32 actorParams = TWO_ACTOR_PARAMS(posX, posZ);
+    Rando::Location* location = OTRGlobals::Instance->gRandomizer->GetCheckObjectFromActor(ACTOR_EN_WOOD02, sceneNum, actorParams);
+    if (location->GetRandomizerCheck() != RC_UNKNOWN_CHECK &&
+        (location->GetRCType() != RCTYPE_NLTREE || OTRGlobals::Instance->gRandomizer->GetRandoSettingValue(RSK_LOGIC_RULES) == RO_LOGIC_NO_LOGIC)) {
+        treeIdentity.randomizerInf = rcToRandomizerInf[location->GetRandomizerCheck()];
+        treeIdentity.randomizerCheck = location->GetRandomizerCheck();
+        return treeIdentity;
+    }
+
+    treeIdentity.randomizerInf = RAND_INF_MAX;
+    treeIdentity.randomizerCheck = RC_UNKNOWN_CHECK;
+    return treeIdentity;
+}
+
 void EnWood02_RandomizerInit(void* actorRef) {
     EnWood02* treeActor = static_cast<EnWood02*>(actorRef);
     if ((treeActor->actor.params <= WOOD_TREE_KAKARIKO_ADULT &&
@@ -132,7 +154,7 @@ void EnWood02_RandomizerInit(void* actorRef) {
         (treeActor->actor.params > WOOD_TREE_KAKARIKO_ADULT &&
          treeActor->actor.params <= WOOD_BUSH_BLACK_LARGE_SPAWNED &&
          Rando::Context::GetInstance()->GetOption(RSK_SHUFFLE_BUSHES).Get())) {
-        auto treeIdentity = OTRGlobals::Instance->gRandomizer->IdentifyTree(
+        auto treeIdentity = IdentifyTree(
             gPlayState->sceneNum, (s16)treeActor->actor.world.pos.x, (s16)treeActor->actor.world.pos.z);
         if (treeIdentity.randomizerInf != RAND_INF_MAX && treeIdentity.randomizerCheck != RC_UNKNOWN_CHECK) {
             ObjectExtension::GetInstance().Set<CheckIdentity>(actorRef, std::move(treeIdentity));
