+++ soh/soh/Enhancements/randomizer/fishsanity.h
@@ -82,7 +82,7 @@ class Fishsanity {
 
     /**
      * @brief Returns the identity for a caught pond fish given its params.
-     * Not for use externally from rando, use Randomizer::IdentifyFish
+     * Not for use externally from rando
      *
      * @param fishParams Actor parameters for the fish to identify
      */
