+++ soh/soh/Enhancements/randomizer/randomizer.cpp
@@ -724,28 +724,6 @@ ItemObtainability Randomizer::GetItemObtainabilityFromRandomizerGet(RandomizerGe
     }
 }
 
-CheckIdentity Randomizer::IdentifyBeehive(s32 sceneNum, s16 xPosition, s32 respawnData) {
-    CheckIdentity beehiveIdentity;
-
-    beehiveIdentity.randomizerInf = RAND_INF_MAX;
-    beehiveIdentity.randomizerCheck = RC_UNKNOWN_CHECK;
-
-    if (sceneNum == SCENE_GROTTOS) {
-        respawnData = TWO_ACTOR_PARAMS(xPosition, respawnData);
-    } else {
-        respawnData = TWO_ACTOR_PARAMS(xPosition, 0);
-    }
-
-    Rando::Location* location = GetCheckObjectFromActor(ACTOR_OBJ_COMB, sceneNum, respawnData);
-
-    if (location->GetRandomizerCheck() != RC_UNKNOWN_CHECK) {
-        beehiveIdentity.randomizerInf = rcToRandomizerInf[location->GetRandomizerCheck()];
-        beehiveIdentity.randomizerCheck = location->GetRandomizerCheck();
-    }
-
-    return beehiveIdentity;
-}
-
 Rando::Location* Randomizer::GetCheckObjectFromActor(s16 actorId, s16 sceneNum, s32 actorParams = 0x00) {
     RandomizerCheck specialRc = RC_UNKNOWN_CHECK;
     // TODO: Migrate these special cases into table, or at least document why they are special
@@ -892,45 +870,7 @@ Rando::Location* Randomizer::GetCheckObjectFromActor(s16 actorId, s16 sceneNum, 
     return Rando::StaticData::GetLocation(RC_UNKNOWN_CHECK);
 }
 
-ScrubIdentity Randomizer::IdentifyScrub(s32 sceneNum, s32 actorParams, s32 respawnData) {
-    struct ScrubIdentity scrubIdentity;
-
-    scrubIdentity.identity.randomizerInf = RAND_INF_MAX;
-    scrubIdentity.identity.randomizerCheck = RC_UNKNOWN_CHECK;
-    scrubIdentity.getItemId = GI_NONE;
-    scrubIdentity.itemPrice = -1;
-
-    // Scrubs that are 0x06 are loaded as 0x03 when child, switching from selling arrows to seeds
-    if (actorParams == 0x06)
-        actorParams = 0x03;
-
-    if (sceneNum == SCENE_GROTTOS) {
-        actorParams = TWO_ACTOR_PARAMS(actorParams, respawnData);
-    }
-
-    Rando::Location* location = GetCheckObjectFromActor(ACTOR_EN_DNS, sceneNum, actorParams);
-
-    if (location->GetRandomizerCheck() != RC_UNKNOWN_CHECK) {
-        if (location->GetRandomizerCheck() == RC_HF_DEKU_SCRUB_GROTTO ||
-            location->GetRandomizerCheck() == RC_LW_DEKU_SCRUB_GROTTO_FRONT ||
-            location->GetRandomizerCheck() == RC_LW_DEKU_SCRUB_NEAR_BRIDGE) {
-            if (GetRandoSettingValue(RSK_SHUFFLE_SCRUBS) == RO_SCRUBS_OFF) {
-                return scrubIdentity;
-            }
-        } else if (GetRandoSettingValue(RSK_SHUFFLE_SCRUBS) != RO_SCRUBS_ALL) {
-            return scrubIdentity;
-        }
-
-        scrubIdentity.identity.randomizerInf = rcToRandomizerInf[location->GetRandomizerCheck()];
-        scrubIdentity.identity.randomizerCheck = location->GetRandomizerCheck();
-        scrubIdentity.getItemId = (GetItemID)Rando::StaticData::RetrieveItem(location->GetVanillaItem()).GetItemID();
-        scrubIdentity.itemPrice =
-            OTRGlobals::Instance->gRandoContext->GetItemLocation(scrubIdentity.identity.randomizerCheck)->GetPrice();
-    }
-
-    return scrubIdentity;
-}
-
+// RANDOTODO: Move all Shopsanity stuff to a ShuffleShops.cpp
 ShopItemIdentity Randomizer::IdentifyShopItem(s32 sceneNum, u8 slotIndex) {
     ShopItemIdentity shopItemIdentity;
 
@@ -967,386 +907,6 @@ ShopItemIdentity Randomizer::IdentifyShopItem(s32 sceneNum, u8 slotIndex) {
     }
 
     return shopItemIdentity;
-}
-
-CheckIdentity Randomizer::IdentifyCow(s32 sceneNum, s32 posX, s32 posZ) {
-    CheckIdentity cowIdentity;
-
-    cowIdentity.randomizerInf = RAND_INF_MAX;
-    cowIdentity.randomizerCheck = RC_UNKNOWN_CHECK;
-
-    s32 actorParams = 0x00;
-    // Only need to pass params if in a scene with two cows
-    if (sceneNum == SCENE_GROTTOS || sceneNum == SCENE_STABLE || sceneNum == SCENE_LON_LON_BUILDINGS) {
-        actorParams = TWO_ACTOR_PARAMS(posX, posZ);
-    }
-
-    Rando::Location* location = GetCheckObjectFromActor(ACTOR_EN_COW, sceneNum, actorParams);
-
-    if (location->GetRandomizerCheck() != RC_UNKNOWN_CHECK) {
-        cowIdentity.randomizerInf = rcToRandomizerInf[location->GetRandomizerCheck()];
-        cowIdentity.randomizerCheck = location->GetRandomizerCheck();
-    }
-
-    return cowIdentity;
-}
-
-CheckIdentity Randomizer::IdentifyPot(s32 sceneNum, s32 posX, s32 posZ) {
-    CheckIdentity potIdentity;
-    uint32_t potSceneNum = sceneNum;
-
-    if (sceneNum == SCENE_GANONDORF_BOSS) {
-        potSceneNum = SCENE_GANONS_TOWER;
-    }
-
-    potIdentity.randomizerInf = RAND_INF_MAX;
-    potIdentity.randomizerCheck = RC_UNKNOWN_CHECK;
-
-    s32 actorParams = TWO_ACTOR_PARAMS(posX, posZ);
-
-    Rando::Location* location = GetCheckObjectFromActor(ACTOR_OBJ_TSUBO, potSceneNum, actorParams);
-
-    if (location->GetRandomizerCheck() == RC_UNKNOWN_CHECK) {
-        LUSLOG_WARN("IdentifyPot did not receive a valid RC value (%d).", location->GetRandomizerCheck());
-    } else {
-        potIdentity.randomizerInf = rcToRandomizerInf[location->GetRandomizerCheck()];
-        potIdentity.randomizerCheck = location->GetRandomizerCheck();
-    }
-
-    return potIdentity;
-}
-
-CheckIdentity Randomizer::IdentifyFish(s32 sceneNum, s32 actorParams) {
-    CheckIdentity fishIdentity;
-
-    fishIdentity.randomizerInf = RAND_INF_MAX;
-    fishIdentity.randomizerCheck = RC_UNKNOWN_CHECK;
-
-    // Fishsanity will determine what the identity of the fish should be
-    if (sceneNum == SCENE_FISHING_POND) {
-        return OTRGlobals::Instance->gRandoContext->GetFishsanity()->IdentifyPondFish(actorParams);
-    }
-
-    Rando::Location* location = GetCheckObjectFromActor(ACTOR_EN_FISH, sceneNum, actorParams);
-
-    if (location->GetRandomizerCheck() != RC_UNKNOWN_CHECK) {
-        fishIdentity.randomizerInf = rcToRandomizerInf[location->GetRandomizerCheck()];
-        fishIdentity.randomizerCheck = location->GetRandomizerCheck();
-    }
-
-    return fishIdentity;
-}
-
-CheckIdentity Randomizer::IdentifyGrass(s32 sceneNum, s32 posX, s32 posZ, s32 respawnData, s32 linkAge) {
-    CheckIdentity grassIdentity;
-
-    grassIdentity.randomizerInf = RAND_INF_MAX;
-    grassIdentity.randomizerCheck = RC_UNKNOWN_CHECK;
-
-    if (sceneNum == SCENE_GROTTOS) {
-        respawnData = TWO_ACTOR_PARAMS(posX, respawnData);
-    } else {
-        // We'll just pretend it's always daytime for our market bushes.
-        if (sceneNum == SCENE_MARKET_NIGHT) {
-            sceneNum = SCENE_MARKET_DAY;
-
-            /*
-                The two bushes by the tree are not in the same spot
-                between night and day. We'll assume the coordinates
-                of the daytime bushes so that we can count them as
-                the same locations.
-            */
-            if (posX == -74) {
-                posX = -106;
-                posZ = 277;
-            }
-            if (posX == -87) {
-                posX = -131;
-                posZ = 225;
-            }
-        }
-
-        /*
-            Same as with Market. ZR has a bush slightly off pos
-            between Child and Adult. This is to merge them into
-            a single location.
-        */
-        if (sceneNum == SCENE_ZORAS_RIVER) {
-            if (posX == 233) {
-                posX = 231;
-                posZ = -1478;
-            }
-        }
-
-        // The two bushes behind the sign in KF should be separate
-        // locations between Child and Adult.
-        if (sceneNum == SCENE_KOKIRI_FOREST && linkAge == 0) {
-            if (posX == -498 || posX == -523) {
-                posZ = 0xFF;
-            }
-        }
-
-        respawnData = TWO_ACTOR_PARAMS(posX, posZ);
-    }
-
-    Rando::Location* location = GetCheckObjectFromActor(ACTOR_EN_KUSA, sceneNum, respawnData);
-
-    if (location->GetRandomizerCheck() != RC_UNKNOWN_CHECK) {
-        grassIdentity.randomizerInf = rcToRandomizerInf[location->GetRandomizerCheck()];
-        grassIdentity.randomizerCheck = location->GetRandomizerCheck();
-    }
-
-    return grassIdentity;
-}
-
-CheckIdentity Randomizer::IdentifyCrate(s32 sceneNum, s32 posX, s32 posZ) {
-    CheckIdentity crateIdentity;
-    uint32_t crateSceneNum = sceneNum;
-
-    // pretend night is day to align crates in market and align GF child/adult crates
-    if (sceneNum == SCENE_MARKET_NIGHT) {
-        crateSceneNum = SCENE_MARKET_DAY;
-    } else if (sceneNum == SCENE_GERUDOS_FORTRESS && gPlayState->linkAgeOnLoad == 1 && posX == 310) {
-        if (posZ == -1830) {
-            posZ = -1842;
-        } else if (posZ == -1770) {
-            posZ = -1782;
-        }
-    }
-
-    crateIdentity.randomizerInf = RAND_INF_MAX;
-    crateIdentity.randomizerCheck = RC_UNKNOWN_CHECK;
-
-    s32 actorParams = TWO_ACTOR_PARAMS(posX, posZ);
-
-    Rando::Location* location = GetCheckObjectFromActor(ACTOR_OBJ_KIBAKO2, crateSceneNum, actorParams);
-
-    if (location->GetRandomizerCheck() == RC_UNKNOWN_CHECK) {
-        LUSLOG_WARN("IdentifyCrate did not receive a valid RC value (%d).", location->GetRandomizerCheck());
-        assert(false);
-    } else {
-        crateIdentity.randomizerInf = rcToRandomizerInf[location->GetRandomizerCheck()];
-        crateIdentity.randomizerCheck = location->GetRandomizerCheck();
-    }
-
-    return crateIdentity;
-}
-
-CheckIdentity Randomizer::IdentifySmallCrate(s32 sceneNum, s32 posX, s32 posZ) {
-    CheckIdentity smallCrateIdentity;
-    uint32_t smallCrateSceneNum = sceneNum;
-
-    smallCrateIdentity.randomizerInf = RAND_INF_MAX;
-    smallCrateIdentity.randomizerCheck = RC_UNKNOWN_CHECK;
-
-    s32 actorParams = TWO_ACTOR_PARAMS(posX, posZ);
-
-    Rando::Location* location = GetCheckObjectFromActor(ACTOR_OBJ_KIBAKO, smallCrateSceneNum, actorParams);
-
-    if (location->GetRandomizerCheck() == RC_UNKNOWN_CHECK) {
-        LUSLOG_WARN("IdentifyCrate did not receive a valid RC value (%d).", location->GetRandomizerCheck());
-        assert(false);
-    } else {
-        smallCrateIdentity.randomizerInf = rcToRandomizerInf[location->GetRandomizerCheck()];
-        smallCrateIdentity.randomizerCheck = location->GetRandomizerCheck();
-    }
-
-    return smallCrateIdentity;
-}
-
-CheckIdentity Randomizer::IdentifyRock(s32 sceneNum, s32 posX, s32 posZ) {
-    CheckIdentity rockIdentity;
-
-    rockIdentity.randomizerInf = RAND_INF_MAX;
-    rockIdentity.randomizerCheck = RC_UNKNOWN_CHECK;
-
-    Rando::Location* location = GetCheckObjectFromActor(ACTOR_EN_ISHI, sceneNum, TWO_ACTOR_PARAMS(posX, posZ));
-
-    if (location->GetRandomizerCheck() != RC_UNKNOWN_CHECK) {
-        rockIdentity.randomizerInf = rcToRandomizerInf[location->GetRandomizerCheck()];
-        rockIdentity.randomizerCheck = location->GetRandomizerCheck();
-    } else {
-        LUSLOG_WARN("IdentifyRock did not receive a valid RC value %d,%d.", posX, posZ);
-    }
-
-    return rockIdentity;
-}
-
-CheckIdentity Randomizer::IdentifyTree(s32 sceneNum, s32 posX, s32 posZ) {
-    CheckIdentity treeIdentity;
-
-    if (sceneNum == SCENE_MARKET_NIGHT) {
-        sceneNum = SCENE_MARKET_DAY;
-    }
-
-    s32 actorParams = TWO_ACTOR_PARAMS(posX, posZ);
-    Rando::Location* location = GetCheckObjectFromActor(ACTOR_EN_WOOD02, sceneNum, actorParams);
-    if (location->GetRandomizerCheck() != RC_UNKNOWN_CHECK &&
-        (location->GetRCType() != RCTYPE_NLTREE || GetRandoSettingValue(RSK_LOGIC_RULES) == RO_LOGIC_NO_LOGIC)) {
-        treeIdentity.randomizerInf = rcToRandomizerInf[location->GetRandomizerCheck()];
-        treeIdentity.randomizerCheck = location->GetRandomizerCheck();
-        return treeIdentity;
-    }
-
-    treeIdentity.randomizerInf = RAND_INF_MAX;
-    treeIdentity.randomizerCheck = RC_UNKNOWN_CHECK;
-    return treeIdentity;
-}
-
-CheckIdentity Randomizer::IdentifySign(s32 sceneNum, s32 posX, s32 posZ, s32 id) {
-    CheckIdentity signIdentity;
-    uint32_t signSceneNum = sceneNum;
-    Rando::Location* location = nullptr;
-
-    // align child/adult signs
-    if (sceneNum == SCENE_KAKARIKO_VILLAGE && LINK_IS_ADULT && posX == 1165 && posZ == 1545) {
-        posZ = 1550;
-    } else if (sceneNum == SCENE_GRAVEYARD && LINK_IS_ADULT) {
-        if (id == ACTOR_EN_WONDER_TALK2 && posX == -807 && posZ == 266) {
-            posX = -805;
-        } else if (id == ACTOR_EN_WONDER_TALK) {
-            if (posX == 634 && posZ == 260) {
-                posX = 654;
-                posZ = 258;
-            } else if (posX == 634 && posZ == -100) {
-                posX = 654;
-                posZ = -102;
-            } else if (posX == 753 && posZ == 85) {
-                posX = 752;
-            }
-        }
-    } else if (sceneNum == SCENE_ZORAS_RIVER && LINK_IS_ADULT && posX == 4097 && posZ == -1399) {
-        posX = 4096;
-        posZ = -1401;
-    }
-
-    signIdentity.randomizerInf = RAND_INF_MAX;
-    signIdentity.randomizerCheck = RC_UNKNOWN_CHECK;
-
-    s32 actorParams = TWO_ACTOR_PARAMS(posX, posZ);
-
-    switch (id) {
-        case ACTOR_EN_KANBAN:
-            location = GetCheckObjectFromActor(ACTOR_EN_KANBAN, signSceneNum, actorParams);
-            break;
-        case ACTOR_EN_A_OBJ:
-            location = GetCheckObjectFromActor(ACTOR_EN_A_OBJ, signSceneNum, actorParams);
-            break;
-        case ACTOR_EN_WONDER_TALK2:
-            location = GetCheckObjectFromActor(ACTOR_EN_WONDER_TALK2, signSceneNum, actorParams);
-            break;
-        case ACTOR_EN_WONDER_TALK:
-            location = GetCheckObjectFromActor(ACTOR_EN_WONDER_TALK, signSceneNum, actorParams);
-            break;
-        default:
-            return signIdentity;
-    }
-
-    if (location == nullptr || location->GetRandomizerCheck() == RC_UNKNOWN_CHECK) {
-        LUSLOG_WARN("IdentifySign did not receive a valid RC value (%d).", location->GetRandomizerCheck());
-    } else {
-        signIdentity.randomizerInf = rcToRandomizerInf[location->GetRandomizerCheck()];
-        signIdentity.randomizerCheck = location->GetRandomizerCheck();
-    }
-
-    return signIdentity;
-}
-
-CheckIdentity Randomizer::IdentifyWonderItem(s32 sceneNum, s32 par1, s32 par2) {
-    CheckIdentity wonderIdentity;
-    uint32_t wonderSceneNum = sceneNum;
-
-    // align oasis trees in colossus between child/adult
-    if (sceneNum == SCENE_DESERT_COLOSSUS && LINK_IS_ADULT) {
-        if (par1 == 1157 && par2 == 2388) {
-            par1 = 1161;
-            par2 = 2383;
-        } else if (par1 == 1114 && par2 == 2580) {
-            par1 = 1113;
-            par2 = 2581;
-        }
-    }
-
-    wonderIdentity.randomizerInf = RAND_INF_MAX;
-    wonderIdentity.randomizerCheck = RC_UNKNOWN_CHECK;
-
-    s32 actorParams = TWO_ACTOR_PARAMS(par1, par2);
-
-    Rando::Location* location = GetCheckObjectFromActor(ACTOR_EN_WONDER_ITEM, wonderSceneNum, actorParams);
-
-    if (location->GetRandomizerCheck() == RC_UNKNOWN_CHECK) {
-        LUSLOG_WARN("IdentifyWonderItem did not receive a valid RC value (%d).", location->GetRandomizerCheck());
-    } else {
-        wonderIdentity.randomizerInf = rcToRandomizerInf[location->GetRandomizerCheck()];
-        wonderIdentity.randomizerCheck = location->GetRandomizerCheck();
-    }
-
-    return wonderIdentity;
-}
-
-CheckIdentity Randomizer::IdentifyBeggar(s32 sceneNum, s32 textId) {
-    CheckIdentity beggarIdentity;
-    beggarIdentity.randomizerInf = RAND_INF_MAX;
-    beggarIdentity.randomizerCheck = RC_UNKNOWN_CHECK;
-
-    Rando::Location* location = GetCheckObjectFromActor(ACTOR_EN_HY, sceneNum, textId);
-    if (location->GetRandomizerCheck() == RC_UNKNOWN_CHECK) {
-        LUSLOG_WARN("IdentifyBeggar did not receive a valid RC value (%d).", location->GetRandomizerCheck());
-    } else {
-        beggarIdentity.randomizerInf = rcToRandomizerInf[location->GetRandomizerCheck()];
-        beggarIdentity.randomizerCheck = location->GetRandomizerCheck();
-    }
-
-    return beggarIdentity;
-}
-
-CheckIdentity Randomizer::IdentifyIcicle(s32 sceneNum, s32 posX, s32 posZ) {
-    struct CheckIdentity icicleIdentity;
-    uint32_t icicleSceneNum = sceneNum;
-
-    icicleIdentity.randomizerInf = RAND_INF_MAX;
-    icicleIdentity.randomizerCheck = RC_UNKNOWN_CHECK;
-
-    s32 actorParams = TWO_ACTOR_PARAMS(posX, posZ);
-
-    Rando::Location* location = GetCheckObjectFromActor(ACTOR_BG_ICE_TURARA, icicleSceneNum, actorParams);
-
-    if (location->GetRandomizerCheck() == RC_UNKNOWN_CHECK) {
-        LUSLOG_WARN("IdentifyIcicle did not receive a valid RC value (%d).", location->GetRandomizerCheck());
-        assert(false);
-    } else {
-        icicleIdentity.randomizerInf = rcToRandomizerInf[location->GetRandomizerCheck()];
-        icicleIdentity.randomizerCheck = location->GetRandomizerCheck();
-    }
-
-    return icicleIdentity;
-}
-
-CheckIdentity Randomizer::IdentifyRedIce(s32 sceneNum, s32 posX, s32 posZ) {
-    struct CheckIdentity redIceIdentity;
-    uint32_t redIceSceneNum = sceneNum;
-
-    // Handle KZ moving
-    if (sceneNum == SCENE_ZORAS_DOMAIN && LINK_IS_ADULT && posX == 531 && posZ == -1818) {
-        posX = 628;
-    }
-
-    redIceIdentity.randomizerInf = RAND_INF_MAX;
-    redIceIdentity.randomizerCheck = RC_UNKNOWN_CHECK;
-
-    s32 actorParams = TWO_ACTOR_PARAMS(posX, posZ);
-
-    Rando::Location* location = GetCheckObjectFromActor(ACTOR_BG_ICE_SHELTER, redIceSceneNum, actorParams);
-
-    if (location->GetRandomizerCheck() == RC_UNKNOWN_CHECK) {
-        LUSLOG_WARN("IdentifyRedIce did not receive a valid RC value (%d).", location->GetRandomizerCheck());
-        assert(false);
-    } else {
-        redIceIdentity.randomizerInf = rcToRandomizerInf[location->GetRandomizerCheck()];
-        redIceIdentity.randomizerCheck = location->GetRandomizerCheck();
-    }
-
-    return redIceIdentity;
 }
 
 u8 Randomizer::GetRandoSettingValue(RandomizerSettingKey randoSettingKey) {
